add doc input `File::class`.
