+++ docs/README.md
@@ -58,6 +58,7 @@ This is particularly useful for creating forms and other interactive elements on
 - [datetime](/docs/form/input/Datetime.md)
 - [datetime-local](/docs/form/input/DatetimeLocal.md)
 - [email](/docs/form/input/Email.md)
+- [file](/docs/form/input/File.md)
 - [radio](/docs/form/input/Radio.md)
 - [radio-list](/docs/form/input/RadioList.md)
 - [text](/docs/form/input/Text.md)
