+++ tests/Input/File/CustomMethodTest.php
@@ -109,6 +109,16 @@ final class CustomMethodTest extends TestCase
         );
     }
 
+    public function testRender(): void
+    {
+        Assert::equalsWithoutLE(
+            <<<HTML
+            <input id="file-65a15e0439570" type="file">
+            HTML,
+            File::widget()->id('file-65a15e0439570')->render()
+        );
+    }
+
     public function testSuffix(): void
     {
         Assert::equalsWithoutLE(
